Move app consts to own package (named app)
