+++ godoc.go
@@ -1,6 +1,7 @@
 // Murex is a cross-platform shell like Bash but with greater emphasis on writing safe shell scripts and powerful one-liners while maintaining readability.
 //
 // A quick breakdown of the project structure:
+//     * app is for hardcoded strings like copyright dates
 //     * builtins contains all the plugins exposed via the languages runtime
 //     * config is APIs for the runtime config
 //     * debug is debugging APIs
