+++ shell/shell.go
@@ -5,8 +5,8 @@ import (
 	"os"
 	"strings"
 
+	"github.com/lmorg/murex/app"
 	"github.com/lmorg/murex/builtins/pipes/term"
-	"github.com/lmorg/murex/config"
 	"github.com/lmorg/murex/lang"
 	"github.com/lmorg/murex/lang/types"
 	"github.com/lmorg/murex/shell/history"
@@ -190,7 +190,7 @@ func ShowPrompt() {
 			merged = ""
 
 			fork := lang.ShellProcess.Fork(lang.F_PARENT_VARTABLE | lang.F_NEW_MODULE | lang.F_NO_STDIN)
-			fork.FileRef.Source.Module = config.AppName
+			fork.FileRef.Source.Module = app.Name
 			fork.Stderr = term.NewErr(ansi.IsAllowed())
 			fork.PromptId = thisProc
 			lang.ShellExitNum, _ = fork.Execute(expanded)
